option to en-/disable jumpcutting

+++ src/downloader.py
@@ -6,7 +6,7 @@ from multiprocessing import Process, Semaphore
 from pathlib import Path
 
 
-def download_list_of_videos(videos: [(str, str)], output_folder_path: Path, tmp_directory: Path, semaphore: Semaphore):
+def download_list_of_videos(videos: [(str, str)], output_folder_path: Path, tmp_directory: Path, semaphore: Semaphore, jumpcut: bool):
     for filename, url in videos:
         filename = re.sub('[\\\\/:*?"<>|]|[\x00-\x20]', '_', filename) + ".mp4"  # Filter illegal filename chars
         output_file_path = Path(output_folder_path, filename)
@@ -16,11 +16,11 @@ def download_list_of_videos(videos: [(str, str)], output_folder_path: Path, tmp_
                 or output_file_path.exists()):  # Check if file exists (we downloaded and converted it already)
             Path(output_file_path.as_posix() + ".lock").touch()  # Create lock file
             Process(target=download_and_cut_video,  # Download video in separate process
-                    args=(filename, url, output_file_path, tmp_directory, semaphore)).start()
+                    args=(filename, url, output_file_path, tmp_directory, semaphore, jumpcut)).start()
 
 
 def download_and_cut_video(filename: str, playlist_url: str, output_file_path: Path, tmp_directory: Path,
-                           semaphore: Semaphore):
+        semaphore: Semaphore, jumpcut: bool):
     semaphore.acquire()  # Acquire lock
 
     temporary_path = Path(tmp_directory, filename + ".original")  # Download location
@@ -34,7 +34,7 @@ def download_and_cut_video(filename: str, playlist_url: str, output_file_path: P
         '-i', playlist_url,  # Input file
         '-c', 'copy',  # Codec name
         '-f', 'mp4',  # Force mp4 as output file format
-        temporary_path  # Output file
+        temporary_path if jumpcut else output_file_path  # Output file
     ], capture_output=True)
 
     if ffmpeg.returncode != 0:  # Print debug output in case of error
@@ -47,30 +47,32 @@ def download_and_cut_video(filename: str, playlist_url: str, output_file_path: P
         return
 
     print(f"Download of {filename} completed after {(time.time() - download_start_time):.0f}s")
-    conversion_start_time = time.time()  # Track auto-editor time
-    print(f"Conversion of {filename} started")
+    if jumpcut:
+        conversion_start_time = time.time()  # Track auto-editor time
+        print(f"Conversion of {filename} started")
 
-    auto_editor = subprocess.run([
-        'auto-editor',
-        temporary_path,  # Input file
-        '--silent_speed', '8',  # Speed multiplier while there is no audio
-        '--video_codec', 'libx264',  # Video codec
-        '--constant_rate_factor', '30',  # Framerate
-        '--no_open',  # Don't open the finished file
-        '-o', output_file_path  # Output file
-    ], capture_output=True)
+        auto_editor = subprocess.run([
+            'auto-editor',
+            temporary_path,  # Input file
+            '--silent_speed', '8',  # Speed multiplier while there is no audio
+            '--video_codec', 'libx264',  # Video codec
+            '--constant_rate_factor', '30',  # Framerate
+            '--no_open',  # Don't open the finished file
+            '-o', output_file_path  # Output file
+        ], capture_output=True)
 
-    if auto_editor.returncode != 0:  # Print debug output in case of error
-        print(f'Error during conversion of "{filename}" with auto-editor:', file=sys.stderr)
-        print(f'Playlist file: {playlist_url}', file=sys.stderr)
-        print(f'Reading from: {temporary_path}', file=sys.stderr)
-        print(f'Designated output location: {output_file_path}', file=sys.stderr)
-        print(f'Output of auto-editor to stdout:\n' + ffmpeg.stdout.decode('utf-8'), file=sys.stderr)
-        print(f'Output of auto-editor to stderr:\n' + ffmpeg.stderr.decode('utf-8'), file=sys.stderr)
-        return
+        if auto_editor.returncode != 0:  # Print debug output in case of error
+            print(f'Error during conversion of "{filename}" with auto-editor:', file=sys.stderr)
+            print(f'Playlist file: {playlist_url}', file=sys.stderr)
+            print(f'Reading from: {temporary_path}', file=sys.stderr)
+            print(f'Designated output location: {output_file_path}', file=sys.stderr)
+            print(f'Output of auto-editor to stdout:\n' + ffmpeg.stdout.decode('utf-8'), file=sys.stderr)
+            print(f'Output of auto-editor to stderr:\n' + ffmpeg.stderr.decode('utf-8'), file=sys.stderr)
+            return
 
-    print(f"Conversion of {filename} completed after {(time.time() - conversion_start_time):.0f}s")
-    temporary_path.unlink()  # Delete original file
+        print(f"Conversion of {filename} completed after {(time.time() - conversion_start_time):.0f}s")
+        temporary_path.unlink()  # Delete original file
+
     Path(output_file_path.as_posix() + ".lock").unlink()  # Remove lock file
     print(f"Completed {filename} after {(time.time() - download_start_time):.0f}s")
